feat(workspace): remove primitives
