+++ book/src/sdk/fpvm-backend.md
@@ -5,7 +5,7 @@
 
 Kona is effectively split into two parts:
 
-- OP Stack state transition logic & types (`kona-derive`, `kona-executor`, `kona-mpt`, `kona-primitives`)
+- OP Stack state transition logic (`kona-derive`, `kona-executor`, `kona-mpt`)
 - {{#template ../../templates/glossary-link.md root=./ ref=fault-proof-vm text=Fault Proof VM}} IO and utilities
   (`kona-common`, `kona-common-proc`, `kona-preimage`)
