+++ crates/derive/src/sources/mod.rs
@@ -4,7 +4,7 @@ mod ethereum;
 pub use ethereum::EthereumDataSource;
 
 mod blobs;
-pub use blobs::BlobSource;
+pub use blobs::{BlobData, BlobSource};
 
 mod calldata;
 pub use calldata::CalldataSource;
