--- crates/primitives/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "kona-primitives"
-description = "Primitive types for kona crates"
-version = "0.0.2"
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# General
-anyhow.workspace = true
-thiserror.workspace = true
-
-# Alloy
-alloy-eips.workspace = true
-alloy-primitives = { workspace = true, features = ["rlp"] }
-
-# `serde` feature dependencies
-serde = { workspace = true, optional = true }
-
-# `online` feature dependencies
-tracing = { workspace = true, optional = true }
-revm = { workspace = true, optional = true }
-sha2 = { workspace = true, optional = true }
-c-kzg = { workspace = true, optional = true }
-
-[dev-dependencies]
-serde_json.workspace = true
-
-[features]
-default = ["serde"]
-serde = ["dep:serde", "alloy-primitives/serde"]
-online = [
-  "dep:c-kzg",
-  "dep:sha2",
-  "dep:revm",
-  "dep:tracing",
-  "revm/default",
-]
