--- crates/primitives/README.md
@@ -1,5 +1,0 @@
-# `kona-primitives`
-
-Primitive types for kona crates.
-
-These types SHOULD be re-exported by downstream kona crates.
