+++ crates/providers-alloy/src/beacon_client.rs
@@ -1,10 +1,8 @@
 //! Contains an online implementation of the `BeaconClient` trait.
 
+use alloy_eips::{eip1898::NumHash, eip4844::BlobTransactionSidecarItem};
+use alloy_primitives::FixedBytes;
 use async_trait::async_trait;
-use kona_primitives::{
-    APIBlobSidecar, APIConfigResponse, APIGenesisResponse, APIGetBlobSidecarsResponse,
-    IndexedBlobHash,
-};
 use reqwest::Client;
 
 /// The config spec engine api method.
@@ -15,6 +13,124 @@ pub(crate) const GENESIS_METHOD: &str = "eth/v1/beacon/genesis";
 
 /// The blob sidecars engine api method prefix.
 pub(crate) const SIDECARS_METHOD_PREFIX: &str = "eth/v1/beacon/blob_sidecars";
+
+/// An API blob sidecar.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct APIBlobSidecar {
+    /// The inner blob sidecar.
+    #[serde(flatten)]
+    pub inner: BlobTransactionSidecarItem,
+    /// The signed block header.
+    #[serde(rename = "signed_block_header")]
+    pub signed_block_header: SignedBeaconBlockHeader,
+    // The inclusion-proof of the blob-sidecar into the beacon-block is ignored,
+    // since we verify blobs by their versioned hashes against the execution-layer block instead.
+}
+
+/// A signed beacon block header.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct SignedBeaconBlockHeader {
+    /// The message.
+    pub message: BeaconBlockHeader,
+    // The signature is ignored, since we verify blobs against EL versioned-hashes
+}
+
+/// A beacon block header.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct BeaconBlockHeader {
+    /// The slot.
+    #[serde(with = "alloy_serde::quantity")]
+    pub slot: u64,
+    /// The proposer index.
+    #[serde(with = "alloy_serde::quantity")]
+    pub proposer_index: u64,
+    /// The parent root.
+    #[serde(rename = "parent_root")]
+    pub parent_root: FixedBytes<32>,
+    /// The state root.
+    #[serde(rename = "state_root")]
+    pub state_root: FixedBytes<32>,
+    /// The body root.
+    #[serde(rename = "body_root")]
+    pub body_root: FixedBytes<32>,
+}
+
+/// An response for fetching blob sidecars.
+#[derive(Debug, Default, serde::Serialize, serde::Deserialize)]
+pub struct APIGetBlobSidecarsResponse {
+    /// The data.
+    pub data: Vec<APIBlobSidecar>,
+}
+
+/// A reduced genesis data.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct ReducedGenesisData {
+    /// The genesis time.
+    #[serde(rename = "genesis_time")]
+    #[serde(with = "alloy_serde::quantity")]
+    pub genesis_time: u64,
+}
+
+/// An API genesis response.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct APIGenesisResponse {
+    /// The data.
+    pub data: ReducedGenesisData,
+}
+
+/// A reduced config data.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct ReducedConfigData {
+    /// The seconds per slot.
+    #[serde(rename = "SECONDS_PER_SLOT")]
+    #[serde(with = "alloy_serde::quantity")]
+    pub seconds_per_slot: u64,
+}
+
+/// An API config response.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct APIConfigResponse {
+    /// The data.
+    pub data: ReducedConfigData,
+}
+
+impl APIConfigResponse {
+    /// Creates a new API config response.
+    pub const fn new(seconds_per_slot: u64) -> Self {
+        Self { data: ReducedConfigData { seconds_per_slot } }
+    }
+}
+
+/// An API version response.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct APIVersionResponse {
+    /// The data.
+    pub data: VersionInformation,
+}
+
+/// Version information.
+#[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct VersionInformation {
+    /// The version.
+    pub version: String,
+}
+
+impl APIGenesisResponse {
+    /// Creates a new API genesis response.
+    pub const fn new(genesis_time: u64) -> Self {
+        Self { data: ReducedGenesisData { genesis_time } }
+    }
+}
+
+impl Clone for APIGetBlobSidecarsResponse {
+    fn clone(&self) -> Self {
+        let mut data = Vec::with_capacity(self.data.len());
+        for sidecar in &self.data {
+            data.push(sidecar.clone());
+        }
+        Self { data }
+    }
+}
 
 /// The [BeaconClient] is a thin wrapper around the Beacon API.
 #[async_trait]
@@ -34,7 +150,7 @@ pub trait BeaconClient {
     async fn beacon_blob_side_cars(
         &self,
         slot: u64,
-        hashes: &[IndexedBlobHash],
+        hashes: &[NumHash],
     ) -> Result<Vec<APIBlobSidecar>, Self::Error>;
 }
 
@@ -103,7 +219,7 @@ impl BeaconClient for OnlineBeaconClient {
     async fn beacon_blob_side_cars(
         &self,
         slot: u64,
-        hashes: &[IndexedBlobHash],
+        hashes: &[NumHash],
     ) -> Result<Vec<APIBlobSidecar>, Self::Error> {
         crate::inc!(PROVIDER_CALLS, &["beacon_client", "beacon_blob_side_cars"]);
         crate::timer!(
@@ -142,7 +258,7 @@ impl BeaconClient for OnlineBeaconClient {
         // Filter the sidecars by the hashes, in-order.
         hashes.iter().for_each(|hash| {
             if let Some(sidecar) =
-                raw_response.data.iter().find(|sidecar| sidecar.inner.index == hash.index as u64)
+                raw_response.data.iter().find(|sidecar| sidecar.inner.index == hash.number)
             {
                 sidecars.push(sidecar.clone());
             }
