+++ crates/providers-alloy/src/test_utils.rs
@@ -1,14 +1,12 @@
 //! Test Utilities for Online Providers
 
+use crate::{APIBlobSidecar, APIConfigResponse, APIGenesisResponse, APIGetBlobSidecarsResponse};
+use alloy_eips::eip1898::NumHash;
 use alloy_node_bindings::{Anvil, AnvilInstance};
 use alloy_provider::{network::Ethereum, ReqwestProvider};
 use alloy_rpc_client::RpcClient;
 use alloy_transport_http::Http;
 use async_trait::async_trait;
-use kona_primitives::{
-    APIBlobSidecar, APIConfigResponse, APIGenesisResponse, APIGetBlobSidecarsResponse,
-    IndexedBlobHash,
-};
 use reqwest::Client;
 
 /// Spawns an Anvil instance and returns a provider and the instance.
@@ -71,7 +69,7 @@ impl crate::BeaconClient for MockBeaconClient {
     async fn beacon_blob_side_cars(
         &self,
         _: u64,
-        _: &[IndexedBlobHash],
+        _: &[NumHash],
     ) -> Result<Vec<APIBlobSidecar>, Self::Error> {
         self.blob_sidecars
             .clone()
